feat: simplify governed project GitOps
