+++ oci-and-db/foundation/operations-advisory/multi-cloud-operating-models/multi-cloud-control-plane/codex-plugins/project-gitops/.codex-plugin/plugin.json
@@ -1,6 +1,6 @@
 {
   "name": "project-gitops",
-  "version": "2.2.0",
+  "version": "3.0.0",
   "description": "Governed project infrastructure and lifecycle changes through the Codex app and pull requests",
   "author": {
     "name": "Oracle Operations Advisory"
