+++ oci-and-db/foundation/operations-advisory/multi-cloud-operating-models/multi-cloud-control-plane/codex-plugins/project-gitops/CHANGELOG.md
@@ -1,0 +1,10 @@
+# Changelog
+
+## 3.0.0
+
+- Move resource schemas and change policy to the catalog contract.
+- Use one contract-driven validator entry point.
+- Split operational guidance into focused references.
+
+Plugin releases track engine and prose changes only. Contract changes ship to
+the catalog `main` branch and do not require a plugin reinstall.
