+++ oci-and-db/foundation/operations-advisory/multi-cloud-operating-models/multi-cloud-control-plane/codex-plugins/project-gitops/skills/project-gitops/scripts/validation/__init__.py
@@ -1,0 +1,1 @@
+"""Focused implementation modules for the Project GitOps validator."""
