+++ oci-and-db/foundation/operations-advisory/multi-cloud-operating-models/multi-cloud-control-plane/docs/reference/security.md
@@ -5,10 +5,10 @@ credentials. Cloud Operations manages repository controls, secrets, state, and
 runner identities. A trusted runner accesses the cloud only after a human merges
 an approved change.
 
-## GitHub Free baseline
+## GitHub plan controls
 
-The supplied release uses one `repository-secrets` profile. GitHub Free private
-repositories cannot access
+MCCP uses one `repository-secrets` profile. GitHub Free private repositories
+cannot access
 [organization secrets or variables](https://docs.github.com/en/actions/how-tos/write-workflows/choose-what-workflows-do/use-secrets#creating-secrets-for-an-organization),
 so each environment uses one optional JSON repository secret:
 
@@ -32,13 +32,20 @@ branch protection. On this profile, approval is procedural: restrict direct
 pushes and administrators, record independent review, verify the current plan
 or check, and then merge. Do not describe this as GitHub-enforced approval.
 
+GitHub Team and GitHub Enterprise Cloud add controls for private repositories
+that Cloud Operations should configure during installation. Use an active
+`main` ruleset that requires a pull request and at least one approving review,
+with no direct-push bypass. The installation runbook includes the plan-specific
+configuration and verification. These controls strengthen the same governed
+request path; they do not change the project workflow.
+
 ## Required controls
 
 - Keep shared and project repositories private.
 - Restrict Platform CI writes to Cloud Operations and review every change to
   its `main` branch.
-- Use the reviewed OCI, Azure, and Google Cloud orchestrator commits and
-  reviewed GitHub Action release tags.
+- Use the approved OCI `release-2.1.4` branch, reviewed Azure and Google Cloud
+  orchestrator commits, and reviewed GitHub Action release tags.
 - Isolate state by organization, project, cloud, environment, and region.
 - Give each runner only the cloud permissions and labels required for its scope.
 - Keep non-production and production on separate runners.
@@ -49,12 +56,11 @@ or check, and then merge. Do not describe this as GitHub-enforced approval.
 - Complete the [environment secret isolation test](verify-secret-isolation.md)
   before the first workload request.
 
-## Additional controls for paid GitHub plans
+## Additional GitHub plan capabilities
 
-These controls are not implemented by this release and require a separately
-tested hardened profile. GitHub Pro and GitHub Team can use protected `main`
-branches and environment secrets or variables in private repositories. They do
-not provide required deployment reviewers for private repositories.
+GitHub Team and GitHub Enterprise Cloud can protect private `main` branches
+and use environment secrets or variables. They do not provide required
+deployment reviewers for private repositories.
 
 GitHub Enterprise Cloud can additionally use required deployment reviewers,
 prevent self-review, and keep environment secrets behind an approval gate in
@@ -66,5 +72,9 @@ current GitHub documentation for
 [protected branches](https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/managing-protected-branches/about-protected-branches),
 [deployment environments](https://docs.github.com/en/actions/reference/workflows-and-actions/deployments-and-environments),
 and [runner groups](https://docs.github.com/en/actions/how-tos/manage-runners/self-hosted-runners/manage-access)
-before designing that profile. Do not add these controls as an untested switch
-to the supplied GitHub Free workflow.
+before configuring customer controls.
+
+Official GitHub Actions referenced by major tag can receive breaking changes.
+Cloud Operations reviews those release changes during Platform CI upgrades.
+The workflows retain their reviewed major tags unless an upgrade requires a
+separately reviewed change.
