--- oci-and-db/foundation/operations-advisory/multi-cloud-operating-models/multi-cloud-control-plane/repository-sources/gitops-templates/resources-catalog/schemas/oci/adb.schema.json
@@ -1,107 +1,0 @@
-{
-  "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "$id": "https://github.com/__CUSTOMER_ORG__/gitops-templates/resources-catalog/schemas/oci/adb.schema.json",
-  "title": "OCI Autonomous Database Serverless project manifest",
-  "description": "Reference-only schema for the supported OCI ADB Serverless project manifest. It is not consumed at runtime.",
-  "x-mccp-contract-status": "reference-only",
-  "x-mccp-runtime-source": "platform-ci governed ADB validator and the pinned OCI Orchestrator Autonomous Database contract",
-  "type": "object",
-  "required": ["autonomous_databases_configuration"],
-  "additionalProperties": false,
-  "properties": {
-    "autonomous_databases_configuration": {
-      "type": "object",
-      "required": ["default_compartment_id", "databases"],
-      "additionalProperties": false,
-      "properties": {
-        "default_compartment_id": {"$ref": "#/$defs/ocidOrPlaceholder"},
-        "databases": {
-          "type": "object",
-          "minProperties": 1,
-          "patternProperties": {
-            "^(?:[A-Z0-9][A-Z0-9_-]{0,127}|__[A-Z][A-Z0-9_]{2,99}__)$": {"$ref": "#/$defs/database"}
-          },
-          "additionalProperties": false
-        }
-      }
-    }
-  },
-  "$defs": {
-    "placeholder": {
-      "type": "string",
-      "pattern": "^__[A-Z][A-Z0-9_]{2,99}__$"
-    },
-    "ocidOrPlaceholder": {
-      "type": "string",
-      "anyOf": [
-        {"pattern": "^ocid1\\.[a-z0-9]+\\.oc1\\.[A-Za-z0-9.-]+\\.[A-Za-z0-9_-]+$"},
-        {"$ref": "#/$defs/placeholder"}
-      ]
-    },
-    "adbNameOrPlaceholder": {
-      "type": "string",
-      "anyOf": [
-        {"pattern": "^[A-Za-z][A-Za-z0-9]{0,29}$"},
-        {"$ref": "#/$defs/placeholder"}
-      ]
-    },
-    "displayNameOrPlaceholder": {
-      "type": "string",
-      "anyOf": [
-        {"pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,254}$"},
-        {"$ref": "#/$defs/placeholder"}
-      ]
-    },
-    "database": {
-      "type": "object",
-      "required": [
-        "db_name",
-        "display_name",
-        "is_dedicated",
-        "ecpu_count",
-        "non_dw_storage_size_in_gbs",
-        "db_workload",
-        "license_model",
-        "enable_cpu_auto_scaling",
-        "enable_storage_auto_scaling",
-        "admin_password",
-        "networking"
-      ],
-      "additionalProperties": false,
-      "properties": {
-        "db_name": {"$ref": "#/$defs/adbNameOrPlaceholder"},
-        "display_name": {"$ref": "#/$defs/displayNameOrPlaceholder"},
-        "is_dedicated": {"const": false},
-        "ecpu_count": {"type": "integer", "minimum": 2, "maximum": 512},
-        "non_dw_storage_size_in_gbs": {"type": "integer", "minimum": 20},
-        "db_workload": {"const": "OLTP"},
-        "license_model": {"enum": ["BRING_YOUR_OWN_LICENSE", "LICENSE_INCLUDED"]},
-        "enable_cpu_auto_scaling": {"type": "boolean"},
-        "enable_storage_auto_scaling": {"type": "boolean"},
-        "admin_password": {"$ref": "#/$defs/placeholder"},
-        "networking": {
-          "type": "object",
-          "required": ["enable_private_endpoint", "subnet_id", "network_security_groups"],
-          "additionalProperties": false,
-          "properties": {
-            "enable_private_endpoint": {"const": true},
-            "subnet_id": {"$ref": "#/$defs/ocidOrPlaceholder"},
-            "network_security_groups": {
-              "type": "array",
-              "minItems": 1,
-              "maxItems": 5,
-              "uniqueItems": true,
-              "items": {
-                "type": "string",
-                "anyOf": [
-                  {"pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,127}$"},
-                  {"$ref": "#/$defs/placeholder"}
-                ]
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-}
