+++ oci-and-db/foundation/operations-advisory/multi-cloud-operating-models/multi-cloud-control-plane/repository-sources/nonprod-project-template/README.md
@@ -8,11 +8,15 @@ Project manifests use `<cloud>/<environment>/<region>/...`. The approved
 foundation references for each environment are recorded in
 `environments/<environment>/environment_information.md`.
 
-Project Teams change only supported manifests through reviewed pull requests
-and configure their review ownership from `.github/CODEOWNERS.template`. Cloud
-Operations owns the handoff, protected workflows, runner configuration, and
-deployment credentials. Terraform state and secret values never belong in this
-repository.
+Project Teams change only supported manifests through reviewed pull requests.
+The Project Team owns review ownership. When it is ready to configure it, copy
+`.github/CODEOWNERS.template` to `.github/CODEOWNERS` and replace
+`__PLATFORM_OWNERS__`, `__DEV_OWNERS__`, `__TEST_OWNERS__`, and
+`__UAT_OWNERS__` with GitHub teams. A template-only file is valid until then.
+On a GitHub Free private repository, CODEOWNERS records intended ownership but
+does not enforce review without branch protection. Cloud Operations owns the
+handoff, protected workflows, runner configuration, and deployment
+credentials. Terraform state and secret values never belong in this repository.
 
 Use the organization's
 [approved catalog](https://github.com/__CUSTOMER_ORG__/gitops-templates) and
