+++ oci-and-db/foundation/operations-advisory/multi-cloud-operating-models/multi-cloud-control-plane/repository-sources/optional-ui/app/config.py
@@ -50,9 +50,6 @@ class Settings(BaseSettings):
         "gcp:europe-west2,europe-west4,us-east4"
     )
 
-    # Small, non-secret MCCP installation configuration.
-    mccp_installation_path: str = "mccp-installation.json"
-
     # Audit behavior
     audit_infra_paths: str = "oci/,azure/,gcp/"
