+++ oci-and-db/foundation/operations-advisory/multi-cloud-operating-models/multi-cloud-control-plane/repository-sources/optional-ui/app/routers/operations.py
@@ -23,7 +23,7 @@ from app.forms import (
 from app.helpers import render_partial, render_repository_state_error
 from app.path_validation import validate_path_segment
 from app.services.dashboard_service import DashboardService
-from app.services.installation_service import load_mccp_installation
+from app.services.project_context import environment_options as project_environment_options
 from app.services.git_service import GitService, RepositoryStateError
 from app.services.operations_service import OperationsService
 
@@ -45,10 +45,7 @@ def _validate_operation_scope(cloud: str, environment: str) -> None:
 
 def _environment_options(project: str) -> list[str]:
     """Return the V2 environments permitted by the selected project layout."""
-    installation = load_mccp_installation(settings.mccp_installation_path)
-    if project.startswith("prod-"):
-        return [installation.project_context(project, "prod").environment]
-    return sorted(installation.nonprod_environments)
+    return project_environment_options(project)
 
 RequiredText = Annotated[str, StringConstraints(min_length=1)]
 OptionalText = str
