+++ oci-and-db/foundation/operations-advisory/multi-cloud-operating-models/multi-cloud-control-plane/repository-sources/optional-ui/app/services/project_service.py
@@ -2,7 +2,7 @@
 import logging
 from app.config import settings
 from app.github import github_client as default_github_client
-from app.services.installation_service import InstallationError, load_mccp_installation
+from app.services.project_context import ProjectContextError, project_context
 
 logger = logging.getLogger(__name__)
 
@@ -15,7 +15,7 @@ class ProjectService:
         self.github = github_client or default_github_client
 
     @staticmethod
-    def _filter_projects(repos: list, installation) -> list[dict]:
+    def _filter_projects(repos: list) -> list[dict]:
         """Return only V2 handed-off project repository names."""
         projects = []
         seen_names = set()
@@ -30,12 +30,12 @@ class ProjectService:
             name_lc = name.lower()
             try:
                 if name_lc.startswith("nonprod-"):
-                    installation.project_context(name, "dev")
+                    project_context(name, "dev")
                 elif name_lc.startswith("prod-"):
-                    installation.project_context(name, "prod")
+                    project_context(name, "prod")
                 else:
                     continue
-            except InstallationError:
+            except ProjectContextError:
                 continue
 
             if name_lc.endswith("-template"):
@@ -58,10 +58,7 @@ class ProjectService:
             repos = (user_repos or []) + (org_repos or [])
 
             logger.info(f"Total repos fetched: {len(repos)}")
-            installation = load_mccp_installation(settings.mccp_installation_path)
-            if self.org != installation.customer_org:
-                return []
-            projects = self._filter_projects(repos, installation)
+            projects = self._filter_projects(repos)
 
             logger.info("Projects visible after filtering: %d", len(projects))
             return projects
